+++ src/main/java/coupon/application/CouponService.java
@@ -1,0 +1,13 @@
+package coupon.application;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+class CouponService {
+
+    public void issue() {
+
+    }
+}
