+++ src/main/java/coupon/domain/CouponCategory.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 enum CouponCategory {
