+++ src/main/java/coupon/domain/CouponIssuableDuration.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import java.time.LocalDate;
 import jakarta.persistence.Column;
