+++ src/main/java/coupon/domain/CouponName.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
