--- src/test/java/coupon/CouponApplicationTests.java
@@ -1,13 +1,0 @@
-package coupon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CouponApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
