+++ src/test/java/coupon/domain/CouponApplicableAmountTest.java
@@ -1,9 +1,10 @@
-package coupon;
+package coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import coupon.CouponApplicableAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
